refactor(config): make config internal
