+++ cmd/sdk.go
@@ -6,7 +6,7 @@ import (
 	"github.com/bndr/gotabulate"
 	"github.com/spf13/cobra"
 
-	"github.com/alex-held/devctl/config"
+	"github.com/alex-held/devctl/internal/config/config"
 	"github.com/alex-held/devctl/pkg/cli"
 )
